Added NoteList and FractalComposerManager.  Updated Note, MidiNote and referencing classes so that the durations and startTimes are given in quarter notes and the Note can be a rest.

+++ FractalComposer/src/Scale.java
@@ -5,5 +5,5 @@
  * @author Myron 
  */
 public interface Scale {
-    MidiNote convertToMidiNote(Note note, long startTime);
+    MidiNote convertToMidiNote(Note note, double startTime);
 }
